+++ crates/oxc_parser/src/lexer/mod.rs
@@ -1,4 +1,5 @@
 #![allow(clippy::unnecessary_safety_comment)]
+#![allow(unsafe_code)]
 
 //! An Ecma-262 Lexer / Tokenizer
 //! Prior Arts: